requirements.txt
apache-airflow-providers-google==10.19.0
google-cloud-storage
google-cloud-bigquery-storage
pyarrow
requests
dbt-bigquery==1.8.2